% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFData.R
\name{as_data_backend.HFData}
\alias{as_data_backend.HFData}
\title{Coerce an HFData Object to an mlr3 Data Backend}
\usage{
\method{as_data_backend}{HFData}(data, ...)
}
\arguments{
\item{data}{(\code{HFData})\cr The \code{HFData} object to convert.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An \code{mlr3} \link[mlr3:DataBackend]{DataBackend}.
}
\description{
Converts an \link{HFData} object into an \code{mlr3} \link[mlr3:DataBackend]{DataBackend},
triggering download and parsing of the underlying dataset if this has not
already happened.
}
