% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_desc.R
\name{download_desc}
\alias{download_desc}
\title{download_desc}
\usage{
download_desc(repo_id, revision = "main", ..., files_metadata = FALSE)
}
\arguments{
\item{repo_id}{repo id of the dataset}

\item{revision}{revision of the dataset, default is "main"}

\item{...}{additional arguments passed to httr::GET}

\item{files_metadata}{whether to include files metadata, default is FALSE}
}
\value{
list containing description and downloads of the dataset
}
\description{
it helps in downloading the description
}
\examples{
download_desc("scikit-learn/iris")
}
