% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{htsk}
\alias{htsk}
\title{Create an mlr3 Task from a Hugging Face dataset
Repository ID of the dataset on the Hugging Face Hub,}
\usage{
htsk(
  repo_id,
  config = NULL,
  file_name = NULL,
  target = NULL,
  primary_key = NULL,
  split = NULL,
  revision = NULL,
  ...
)
}
\arguments{
\item{repo_id}{(\code{character(1)})\cr}

\item{config}{(\code{character(1)})\cr}

\item{file_name}{(\code{character(1)})\cr}

\item{target}{(\code{character})\cr}

\item{primary_key}{(\code{character})\cr}

\item{split}{(\code{character(1)})\cr}

\item{revision}{(\code{character(1)})\cr}

\item{...}{Additional arguments passed to \code{\link{HFData}} and \code{\link[mlr3:as_task]{mlr3::as_task()}}.}
}
\value{
An \code{\link[mlr3:Task]{mlr3::Task}} object.
}
\description{
Create an mlr3 Task from a Hugging Face dataset
Repository ID of the dataset on the Hugging Face Hub,
}
