% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_hfhub.R
\name{cache_hfhub}
\alias{cache_hfhub}
\title{cache_hfhub}
\usage{
cache_hfhub(
  repo_id,
  file_name = NULL,
  local_files_only = FALSE,
  revision = "main",
  ...
)
}
\arguments{
\item{repo_id}{repo id of the dataset}

\item{file_name}{file name e.g. "iris.csv"}

\item{local_files_only}{can if data is already present in the disk then that can be accessed}

\item{revision}{it helps in getting any commit_hash, branch datasets}

\item{...}{additional arguments passed to httr::GET}
}
\value{
it returns the path of the dataset were its downloaded
}
\description{
it helps in caching the canonical dataset means the original dataset that are on hf.
}
\examples{
\dontrun{
  cache_hfhub(
    repo_id   = "scikit-learn/iris",
    file_name = "iris.csv",
    fun       = download_hub
  )
}
}
