% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_parquet.R
\name{cache_parquet}
\alias{cache_parquet}
\title{cache_parquet}
\usage{
cache_parquet(repo_id, config = NULL, split = NULL, ...)
}
\arguments{
\item{repo_id}{repo id of the dataset}

\item{config}{if dataset have no config give config as default or you can check the repository config is mainly a folder}

\item{split}{user have option to select a specific split}

\item{...}{additional arguments for future}
}
\value{
list containing split and the url of the split
}
\description{
it helps in caching the dataset which in the branch refs/convert/parquet
}
\examples{
\dontrun{
  cache_parquet(
    repo_id   = "scikit-learn/iris",
    config="default"
  )
}
cache_parquet(repo_id="scikit-learn/iris",config="default")
}
