% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{htsk}
\alias{htsk}
\title{Create an HFData object from a Hugging Face dataset}
\usage{
htsk(
  repo_id,
  config = NULL,
  file_name = NULL,
  target = NULL,
  primary_key = NULL,
  split = NULL,
  ...
)
}
\arguments{
\item{repo_id}{(\code{character(1)})\cr
Repository ID of the dataset on the Hugging Face Hub,
e.g. \code{"scikit-learn/iris"} or \code{"ibm-research/duorc"}.}

\item{config}{(\code{character(1)})\cr
Dataset configuration to load.}

\item{file_name}{(\code{character(1)})\cr
Path to a specific dataset file within the repository. This can be used
instead of \code{config} to load a particular file.}

\item{target}{(\code{character})\cr
Name(s) of the target column(s). If \code{NULL}, the target must be specified
later before converting to an mlr3 task.}

\item{primary_key}{(\code{character})\cr
Name of the column that uniquely identify each observation.}

\item{split}{(\code{character(1)})\cr
Dataset split to load, such as \code{"train"}, \code{"test"}, or \code{"validation"}.
If \code{NULL}, all available splits are loaded.}

\item{...}{Additional arguments passed to \code{\link{HFData}}.}
}
\value{
An \code{\link{HFData}} object.
}
\description{
Creates an \code{\link{HFData}} object for a dataset hosted on the Hugging Face Hub.
The returned object can be converted to an mlr3 task using
\code{\link[mlr3:as_task]{mlr3::as_task()}} or accessed through its active bindings.
}
