% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3hf}
\alias{mlr3hf-package}
\alias{mlr3hf}
\title{mlr3hf: Hugging Face datasets for mlr3}
\description{
Provides integration between Hugging Face datasets and the
\pkg{mlr3} ecosystem. Datasets can be downloaded from the
Hugging Face Hub and converted directly into
\link[mlr3:Task]{mlr3 Tasks}.
}
\section{mlr3 Integration}{


The package registers the \code{"hf"} task in
\code{mlr3::mlr_tasks}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{task = tsk(
  "hf",
  repo_id = "scikit-learn/iris",
  config = "default",
  target = "species"
)
}\if{html}{\out{</div>}}

Alternatively, users can create an \code{HFData} object and convert it
using \code{as_task()}.
}

\section{Caching}{


Downloaded datasets are cached locally to avoid repeated downloads.
}

\section{Documentation}{


See the package vignettes for examples of loading datasets,
converting them to mlr3 tasks, and working with large datasets.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://anjaninandan001.github.io/mlr3hf/}
}

}
\author{
\strong{Maintainer}: Sebastian Fischer \email{sebastian.fischer@stat.uni-muenchen.de}

Authors:
\itemize{
  \item Toby Hocking \email{toby.hocking@r-project.org}
  \item Anjani Nandan \email{anjaninandan599@gmail.com}
}

}
